tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/027.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: doc write added script with content, .src set later</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>
	<script>log('inline script #1');
		document.write('<script>log(\'doc.write script\')<\/script>');
		log('end script #1');
	</script>

	<script type="text/javascript">
	log( 'inline script #2' );
	document.getElementsByTagName('script')[4].src='scripts/include-1.js';
	var t = async_test()

	function test() {
		assert_array_equals(eventOrder, ['inline script #1', 'doc.write script', 'end script #1', 'inline script #2']);
		t.done();
}
	onload = t.step_func(function(){
		setTimeout(t.step_func(test), 100);
	})
	</script>

</body></html>
